+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### ✅ Added
 - Support for custom message receipt states
+- Scrolling of instant commands
 
 ### 🔄 Changed
 - Updated Nuke dependency to 11.3.0 for SPM